refactor to use new signedUrl
